bank-chatbot/requirements.txt
aiobotocore==2.12.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
billiard==4.2.0
botocore==1.34.51
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.5
curl_cffi==0.6.2
dataclasses-json==0.6.4
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
duckduckgo_search==5.2.2
dulwich==0.21.7
dvc==3.43.1
dvc-data==3.9.0
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==3.0.6
dvc-render==1.0.1
dvc-studio-client==0.20.0
dvc-task==0.3.0
entrypoints==0.4
fastapi==0.110.0
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gcsfs==2023.12.2.post1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grandalf==0.8
greenlet==3.0.3
gto==1.6.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.6
iterative-telemetry==0.0.8
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain==0.1.9
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.7
langchain-pinecone==0.0.3
langchainhub==0.1.15
langsmith==0.1.27
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.14.1
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pinecone-client==3.0.0
pinecone-datasets==0.7.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pyarrow==11.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2023.12.2
scikit-learn==1.4.1.post1
scipy==1.12.0
scmrepo==2.1.1
semver==3.0.2
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sqltrie==0.11.0
starlette==0.36.3
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.9.0
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.28.0
uvloop==0.19.0
vine==5.1.0
voluptuous==0.14.2
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
python-multipart

